% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{dialektversum_deps}
\alias{dialektversum_deps}
\title{List all dialektversum dependencies}
\usage{
dialektversum_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
dialektversum packages.}
}
\description{
List all dialektversum dependencies
}
