% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{dialektversum_packages}
\alias{dialektversum_packages}
\title{List all packages in the dialektversum}
\usage{
dialektversum_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include dialektversum in the list?}
}
\description{
List all packages in the dialektversum
}
\examples{
dialektversum_packages()
}
