% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{dialektversum_update}
\alias{dialektversum_update}
\title{Update dialektversum packages}
\usage{
dialektversum_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
dialektversum packages.}
}
\description{
This will check to see if all dialektversum packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
dialektversum_update()
}
}
